fixes permissions group foreignkey

+++ database/migrations/2023_05_10_185138_create_groups_table.php
@@ -15,7 +15,7 @@ class CreateGroupsTable extends Migration
     {
         Schema::create('groups', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('parent_id')->nullable();
+            $table->unsignedBigInteger('parent_id')->nullable();
             $table->string('name');
             $table->smallInteger('sort')->default(0);
             $table->timestamps();
